The great refactor!
The team is attempting to employ some coding standards/practices, including the idea of trying to avoid putting code in the files that CubeMX will generate (Core/Inc and Core/Src), and instead, creating a User directory to store user code.

This project now follows this guideline with the exception of thread and queue definitions still living in Core/Src/app_freertos.c.

I have also added some documentation for the repo as a part of this commit.